+++ core/amber/src/main/scala/texera/common/operators/TexeraSourceOperatorExecutor.scala
@@ -1,11 +1,12 @@
-package texera.common.workflow
+package texera.common.operators
 
 import Engine.Common.SourceOperatorExecutor
+import Engine.Common.tuple.Tuple
 import Engine.Common.tuple.texera.TexeraTuple
 
 trait TexeraSourceOperatorExecutor extends SourceOperatorExecutor {
 
-  override def produce(): Iterator[TexeraTuple] = {
+  override def produce(): Iterator[Tuple] = {
     produceTexeraTuple()
   }
