+++ core/amber/src/main/scala/texera/common/operators/filter/TexeraFilterOpDesc.scala
@@ -2,7 +2,7 @@ package texera.common.operators.filter
 
 import Engine.Common.tuple.texera.schema.Schema
 import com.google.common.base.Preconditions
-import texera.common.workflow.TexeraOperatorDescriptor
+import texera.common.operators.TexeraOperatorDescriptor
 
 abstract class TexeraFilterOpDesc extends TexeraOperatorDescriptor {
