+++ core/amber/src/main/scala/texera/common/operators/flatmap/TexeraFlatMapOpDesc.scala
@@ -1,6 +1,6 @@
 package texera.common.operators.flatmap
 
-import texera.common.workflow.TexeraOperatorDescriptor
+import texera.common.operators.TexeraOperatorDescriptor
 
 abstract class TexeraFlatMapOpDesc extends TexeraOperatorDescriptor {
