+++ core/amber/src/main/scala/texera/common/operators/map/TexeraMapOpDesc.scala
@@ -1,6 +1,6 @@
 package texera.common.operators.map
 
-import texera.common.workflow.TexeraOperatorDescriptor
+import texera.common.operators.TexeraOperatorDescriptor
 
 abstract class TexeraMapOpDesc extends TexeraOperatorDescriptor {
