+++ core/amber/src/main/scala/texera/common/workflow/TexeraWorkflow.scala
@@ -1,5 +1,7 @@
 package texera.common.workflow
 
+import texera.common.operators.TexeraOperatorDescriptor
+
 import scala.collection.mutable
 
 case class TexeraBreakpointInfo(operatorID: String, breakpoint: TexeraBreakpoint)
