+++ core/amber/src/main/scala/texera/common/workflow/TexeraWorkflowCompiler.scala
@@ -7,6 +7,7 @@ import Engine.Common.AmberTag.OperatorTag
 import Engine.Common.tuple.texera.TexeraTuple
 import Engine.Operators.OpExecConfig
 import akka.actor.ActorRef
+import texera.common.operators.TexeraOperatorDescriptor
 import texera.common.{TexeraConstraintViolation, TexeraContext}
 
 import scala.collection.mutable
