+++ core/amber/src/main/scala/texera/operators/localscan/LocalCsvFileScanOpDesc.java
@@ -11,7 +11,7 @@ import org.apache.curator.shaded.com.google.common.io.Files;
 import scala.collection.Seq;
 import texera.common.metadata.OperatorGroupConstants;
 import texera.common.metadata.TexeraOperatorInfo;
-import texera.common.workflow.TexeraOperatorDescriptor;
+import texera.common.operators.TexeraOperatorDescriptor;
 
 import java.io.File;
 import java.io.IOException;
