+++ core/amber/src/main/scala/texera/operators/sink/SimpleSinkOpDesc.java
@@ -6,7 +6,7 @@ import com.google.common.base.Preconditions;
 import scala.collection.Seq;
 import texera.common.metadata.OperatorGroupConstants;
 import texera.common.metadata.TexeraOperatorInfo;
-import texera.common.workflow.TexeraOperatorDescriptor;
+import texera.common.operators.TexeraOperatorDescriptor;
 
 public class SimpleSinkOpDesc extends TexeraOperatorDescriptor {
