+++ core/amber/src/main/scala/web/model/request/ModifyLogicRequest.scala
@@ -1,6 +1,6 @@
 package web.model.request
 
-import texera.common.workflow.TexeraOperatorDescriptor
+import texera.common.operators.TexeraOperatorDescriptor
 
 case class ModifyLogicRequest(
     operator: TexeraOperatorDescriptor
